+++ R/filter_Datetime.R
@@ -32,6 +32,7 @@
 #'   is FALSE). This is useful, e.g., when the first observation in the dataset
 #'   is slightly after midnight. If TRUE, it will count the length from midnight
 #'   on to avoid empty days in plotting with [gg_day()].
+#' @param only_Id An expression of `ids` where the filtering should be applied to. If `NULL` (the default), the filtering will be applied to all `ids`. Based on the this expression, the dataset will be split in two and only where the given expression evaluates to `TRUE`, will the filtering take place. Afterwards both sets are recombined and sorted by `Datetime`.
 #'
 #' @return a `data.frame` object identical to `dataset` but with only the
 #'   specified Dates/Times.
@@ -86,11 +87,14 @@ filter_Datetime <- function(dataset,
                             length = NULL,
                             full.day = FALSE,
                             tz = NULL,
+                            only_Id = NULL,
                             filter.expr = NULL) {
   
   # Initial Checks ----------------------------------------------------------
   
   filter.expr <- rlang::enexpr(filter.expr)
+  
+  only_Id <- rlang::enexpr(only_Id)
   
   Datetime.colname.defused <- 
     rlang::enexpr(Datetime.colname) %>% rlang::as_string()
@@ -119,6 +123,16 @@ filter_Datetime <- function(dataset,
  
   # Manipulation ----------------------------------------------------------
    
+  #split the dataset in two parts, based on the only_Id expression
+  if(!is.null(only_Id)) {
+    dataset_unfiltered <-
+      dataset %>% 
+      dplyr::filter(!(!!only_Id), .preserve = TRUE)
+    dataset <- 
+      dataset %>% 
+      dplyr::filter(!!only_Id, .preserve = TRUE)
+  }
+  
   #calculate starting time if length and end are given
   if(is.null(start) & !is.null(length) & !is.null(end)) {
     start <- lubridate::as_datetime(end, tz = tz) - length
@@ -155,7 +169,10 @@ filter_Datetime <- function(dataset,
     }
   
   # Return --------------------------------------------------------------
-  dataset
+    if(!is.null(only_Id)) {
+      dplyr::bind_rows(dataset, dataset_unfiltered) %>% 
+        dplyr::arrange({{ Datetime.colname }}, .by_group = TRUE)
+    } else dataset
 }
 
 
@@ -184,4 +201,44 @@ filter_Date <- function(...,
   filter_Datetime(...,
                   start = start,
                   end = end)
+}
+
+# multiple filter_Date -------------------------------------------------------------
+
+#' Filter multiple times based on a list of arguments.
+#'
+#' [filter_Datetime_multiple()] is a wrapper around [filter_Datetime()] or
+#' [filter_Date()] that allows the cumulative filtering of `Datetimes` based on
+#' varying filter conditions. It is most useful in conjunction with the
+#' `only_Id` argument, e.g., to selectively cut off dates depending on
+#' participants (see examples)
+#'
+#' @param dataset A light logger dataset
+#' @param arguments A list of arguments to be passed to [filter_Datetime()] or
+#'   [filter_Date()]. each list entry must itself be a list of arguments, e.g,
+#'   `list(start = "2021-01-01", only_Id = quote(Id == 216))`. Expressions have
+#'   to be quoted with [quote()] or [rlang::expr()].
+#' @param filter_function The function to be used for filtering, either
+#'   `filter_Datetime` (the default) or `filter_Date`
+#'
+#' @return A dataframe with the filtered data
+#' @export
+#'
+#' @examples
+#' arguments <- list(
+#'  list(start = "2023-08-17", only_Id = quote(Source == "Participant")),
+#'  list(end = "2023-08-17", only_Id = quote(Source == "Environment")))
+#'  #compare the unfiltered dataset
+#'  sample.data.environment %>% gg_overview(Id.colname = Source)
+#'  #compare the unfiltered dataset
+#'  sample.data.environment %>% 
+#'  filter_Datetime_multiple(arguments = arguments, filter_Date) %>%
+#'  gg_overview(Id.colname = Source)
+filter_Datetime_multiple <- function(dataset, 
+                                     arguments, 
+                                     filter_function = filter_Datetime) {
+  
+  purrr::reduce(arguments, function(dataset, params) {
+    do.call({{ filter_function }}, c(list(dataset = dataset), params))
+  }, .init = dataset)
 }