+++ R/import_LL.R
@@ -47,7 +47,7 @@
 #'   by the `device.ext` spec to access the sample file):
 #' * `"ActLumus"` (ActLumus.txt)
 #' * `"LYS"` (LYS.csv)
-#' * `"Actiwatch_Spectrum"` (Actiwatch.csv) *Note: as the `locale` argument use `readr::locale(encoding="latin1")` . This is due to the fact that the German Actiwatch software from which this sample file was taken, uses a different encoding than UTF-8.*
+#' * `"Actiwatch_Spectrum"` (Actiwatch.csv) **Special Argument: `column_names`** A character vector containing column names in the order in which they appear in the file. This is necessary to find the starting point of actual data.*Note: as the `locale` argument use `readr::locale(encoding="latin1")` . This is due to the fact that the German Actiwatch software from which this sample file was taken, uses a different encoding than UTF-8.*
 #' @importFrom rlang :=
 #' @return Tibble/Dataframe with a POSIXct column for the datetime
 #' @export
@@ -124,6 +124,7 @@ imports <- function(device,
       auto.id = ".*",
       manual.id = NULL,
       locale = readr::default_locale(),
+      silent = FALSE,
       ... =
     ),
     #function expression
@@ -169,7 +170,7 @@ imports <- function(device,
         dplyr::arrange(Datetime, .by_group = TRUE)
       
       #give info about the file
-      import.info(tmp, !!device, tz, !!ID.colname)
+      if(!silent) import.info(tmp, !!device, tz, !!ID.colname)
       
       #return the file
       tmp